corpora preprocessing and feature computation/requirements.txt
anytree==2.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
codetiming==1.4.0
contourpy==1.1.0
cycler==0.11.0
emoji==2.7.0
filelock==3.12.2
fonttools==4.41.1
funcy==2.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lexicalrichness==0.5.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pluggy==1.2.0
protobuf==4.23.4
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
stanza==1.5.0
sympy==1.12
textblob==0.17.1
threadpoolctl==3.2.0
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wordcloud==1.9.2
XlsxWriter==3.1.2
